requirements.txt
pyserial>=3.5
numpy>=1.21.0
matplotlib>=3.5.0
requests>=2.28.0
pyyaml>=6.0